decoupled class Lattice into separate file
